REFACTOR: Use generic method array
