+++ lib/UtapiClient.js
@@ -1,46 +1,10 @@
 const assert = require('assert');
 const { Logger } = require('werelogs');
+const { methods } = require('../utils/genericOperations');
 const Datastore = require('./Datastore');
 const { generateKey, generateCounter, generateStateKey } = require('./schema');
 const { errors } = require('arsenal');
 const redisClient = require('../utils/redisClient');
-
-const methods = {
-    createBucket: '_genericPushMetric',
-    deleteBucket: '_genericPushMetric',
-    listBucket: '_genericPushMetric',
-    getBucketAcl: '_genericPushMetric',
-    putBucketAcl: '_genericPushMetric',
-    putBucketCors: '_genericPushMetric',
-    getBucketCors: '_genericPushMetric',
-    deleteBucketCors: '_genericPushMetric',
-    putBucketWebsite: '_genericPushMetric',
-    getBucketWebsite: '_genericPushMetric',
-    getBucketLocation: '_genericPushMetric',
-    deleteBucketWebsite: '_genericPushMetric',
-    uploadPart: '_pushMetricUploadPart',
-    initiateMultipartUpload: '_genericPushMetric',
-    completeMultipartUpload: '_pushMetricCompleteMultipartUpload',
-    listMultipartUploads: '_pushMetricListBucketMultipartUploads',
-    listMultipartUploadParts: '_genericPushMetric',
-    abortMultipartUpload: '_genericPushMetricDeleteObject',
-    deleteObject: '_genericPushMetricDeleteObject',
-    multiObjectDelete: '_genericPushMetricDeleteObject',
-    getObject: '_pushMetricGetObject',
-    getObjectAcl: '_genericPushMetric',
-    getObjectTagging: '_genericPushMetric',
-    putObject: '_genericPushMetricPutObject',
-    copyObject: '_genericPushMetricPutObject',
-    putObjectAcl: '_genericPushMetric',
-    putObjectTagging: '_genericPushMetric',
-    deleteObjectTagging: '_genericPushMetric',
-    headBucket: '_genericPushMetric',
-    headObject: '_genericPushMetric',
-    putBucketVersioning: '_genericPushMetric',
-    getBucketVersioning: '_genericPushMetric',
-    putDeleteMarkerObject: '_pushMetricDeleteMarkerObject',
-    putBucketReplication: '_genericPushMetric',
-};
 
 const metricObj = {
     buckets: 'bucket',
