+++ lib/schema.js
@@ -1,3 +1,6 @@
+const { genericOperations, getS3Operation } =
+    require('../utils/genericOperations');
+
 // metric type schema
 const stateKeys = {
     storageUtilized: prefix => `${prefix}storageUtilized`,
@@ -9,43 +12,33 @@ const counters = {
     numberOfObjectsCounter: prefix => `${prefix}numberOfObjects:counter`,
 };
 
-const keys = {
-    createBucket: prefix => `${prefix}CreateBucket`,
-    deleteBucket: prefix => `${prefix}DeleteBucket`,
-    listBucket: prefix => `${prefix}ListBucket`,
-    getBucketAcl: prefix => `${prefix}GetBucketAcl`,
-    putBucketAcl: prefix => `${prefix}PutBucketAcl`,
-    putBucketCors: prefix => `${prefix}PutBucketCors`,
-    deleteBucketCors: prefix => `${prefix}DeleteBucketCors`,
-    getBucketCors: prefix => `${prefix}GetBucketCors`,
-    putBucketWebsite: prefix => `${prefix}PutBucketWebsite`,
-    putBucketVersioning: prefix => `${prefix}PutBucketVersioning`,
-    getBucketVersioning: prefix => `${prefix}GetBucketVersioning`,
-    deleteBucketWebsite: prefix => `${prefix}DeleteBucketWebsite`,
-    getBucketWebsite: prefix => `${prefix}GetBucketWebsite`,
-    getBucketLocation: prefix => `${prefix}GetBucketLocation`,
+const schema = {
     listBucketMultipartUploads: prefix => `${prefix}ListBucketMultipartUploads`,
-    listMultipartUploadParts: prefix => `${prefix}ListMultipartUploadParts`,
-    initiateMultipartUpload: prefix => `${prefix}InitiateMultipartUpload`,
     completeMultipartUpload: prefix => `${prefix}CompleteMultipartUpload`,
     abortMultipartUpload: prefix => `${prefix}AbortMultipartUpload`,
     deleteObject: prefix => `${prefix}DeleteObject`,
     multiObjectDelete: prefix => `${prefix}MultiObjectDelete`,
     uploadPart: prefix => `${prefix}UploadPart`,
     getObject: prefix => `${prefix}GetObject`,
-    getObjectAcl: prefix => `${prefix}GetObjectAcl`,
-    getObjectTagging: prefix => `${prefix}GetObjectTagging`,
     putObject: prefix => `${prefix}PutObject`,
     copyObject: prefix => `${prefix}CopyObject`,
-    putObjectAcl: prefix => `${prefix}PutObjectAcl`,
-    putObjectTagging: prefix => `${prefix}PutObjectTagging`,
-    deleteObjectTagging: prefix => `${prefix}DeleteObjectTagging`,
-    headBucket: prefix => `${prefix}HeadBucket`,
-    headObject: prefix => `${prefix}HeadObject`,
-    putBucketReplication: prefix => `${prefix}PutBucketReplication`,
     incomingBytes: prefix => `${prefix}incomingBytes`,
     outgoingBytes: prefix => `${prefix}outgoingBytes`,
 };
+
+/**
+* Get the schema that includes generic operations
+* @return {object} - The schema object with all operations
+*/
+function getCompleteSchema() {
+    const obj = {};
+    genericOperations.forEach(operation => {
+        obj[operation] = prefix => `${prefix}${getS3Operation(operation)}`;
+    });
+    return Object.assign({}, schema, obj);
+}
+
+const keys = getCompleteSchema();
 
 /**
 * Creates the appropriate prefix for schema keys
